+++ CommandLine/src/Worms/Resources/Replays/ReplayTextPrinter.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Worms.Armageddon.Resources.Replays;
 using Worms.Logging.TableOutput;
 
-namespace Worms.Resources.Replays.Text
+namespace Worms.Resources.Replays
 {
     public class ReplayTextPrinter : IResourcePrinter<ReplayResource>
     {
