+++ CommandLine/src/Worms/Resources/Schemes/SchemeTextPrinter.cs
@@ -5,10 +5,9 @@ using System.Linq;
 using Syroot.Worms.Armageddon;
 using Worms.Armageddon.Resources.Schemes;
 using Worms.Armageddon.Resources.Schemes.Text;
-using Worms.Logging;
 using Worms.Logging.TableOutput;
 
-namespace Worms.Resources.Schemes.Text
+namespace Worms.Resources.Schemes
 {
     internal class SchemeTextPrinter : IResourcePrinter<SchemeResource>
     {
